# expense-management-app/API.yaml
openapi: 3.0.0
info:
  title: Expense Management API
  version: 1.0.0
  description: Full-stack expense management API with payment processing and AI insights
  contact:
    name: Support
    url: https://github.com/yourusername/expense-management-app
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.example.com/api
    description: Production server
paths:
...
  /expenses:
    get:
      summary: Get all expenses for current user
      tags: [Expenses]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, reimbursed]
        - name: category
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of expenses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseList'

    post:
      summary: Create new expense with optional receipt
      tags: [Expenses]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                amount:
                  type: number
                  format: decimal
                category_id:
                  type: string
                  format: uuid
                expense_date:
                  type: string
                  format: date
                payment_method:
                  type: string
                  enum: [cash, card, check, online]
                notes:
                  type: string
                receipt:
                  type: string
                  format: binary
      responses:
        '201':
          description: Expense created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
  /expenses/{id}:
    get:
      summary: Get single expense
      tags: [Expenses]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Expense details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'

    put:
      summary: Update expense
      tags: [Expenses]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseInput'
      responses:
        '200':
          description: Expense updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'

    delete:
      summary: Delete expense
      tags: [Expenses]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Expense deleted
  /payments/create-intent:
    post:
      summary: Create Stripe payment intent
      tags: [Payments]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                expense_id:
                  type: string
                  format: uuid
                amount:
                  type: number
                  format: decimal
                currency:
                  type: string
                  default: usd
      responses:
        '201':
          description: Payment intent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntent'
  /payments/confirm:
    post:
      summary: Confirm payment
      tags: [Payments]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                payment_intent_id:
                  type: string
                payment_method_id:
                  type: string
      responses:
        '200':
          description: Payment confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
...
components:
  schemas:
...
    Expense:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        amount:
          type: number
          format: decimal
        category_id:
          type: string
          format: uuid
        category_name:
          type: string
        expense_date:
          type: string
          format: date
        status:
          type: string
          enum: [pending, approved, rejected, reimbursed]
        receipt_url:
          type: string
        payment_method:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ExpenseInput:
      type: object
      required:
        - title
        - amount
        - expense_date
      properties:
        title:
          type: string
        description:
          type: string
        amount:
          type: number
          format: decimal
        category_id:
          type: string
          format: uuid
        expense_date:
          type: string
          format: date
        payment_method:
          type: string
    ExpenseList:
      type: object
      properties:
        expenses:
          type: array
          items:
            $ref: '#/components/schemas/Expense'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        expense_id:
          type: string
          format: uuid
        stripe_payment_id:
          type: string
        amount:
          type: number
          format: decimal
        status:
          type: string
          enum: [pending, processing, succeeded, failed, refunded]
        created_at:
          type: string
          format: date-time
    PaymentIntent:
      type: object
      properties:
        clientSecret:
          type: string
        paymentIntentId:
          type: string
        payment:
          $ref: '#/components/schemas/Payment'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT